backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.23
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.25.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
numpy>=1.26.0
pandas>=2.1.4
python-dotenv>=1.0.0
aiofiles>=23.2.1
Pillow>=10.1.0
reportlab>=4.0.7
qrcode[pil]>=7.4.2
cryptography>=41.0.8
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
openpyxl>=3.1.2
aiosqlite>=0.19.0